Implemented support for Swordfish IV

+++ src/com/maxprograms/remotetm/TmManager.java
@@ -23,6 +23,7 @@ import java.io.IOException;
 import java.nio.file.Files;
 import java.sql.SQLException;
 import java.util.Iterator;
+import java.util.List;
 import java.util.Map;
 import java.util.Set;
 import java.util.concurrent.ConcurrentHashMap;
@@ -31,12 +32,15 @@ import javax.xml.parsers.ParserConfigurationException;
 
 import com.maxprograms.remotetm.utils.Utils;
 import com.maxprograms.swordfish.tm.InternalDatabase;
+import com.maxprograms.swordfish.tm.Match;
+import com.maxprograms.xml.Element;
 
 import org.xml.sax.SAXException;
 
 public class TmManager {
 
     static final String MEMORIES = "memories";
+    static final String CLOSED = "Memory is closed";
 
     private TmManager() {
         // private for security
@@ -47,13 +51,10 @@ public class TmManager {
 
     public static int storeTMX(String memory, String tmx, String project, String client, String subject)
             throws IOException, SQLException, SAXException, ParserConfigurationException {
-        if (databases == null) {
-            databases = new ConcurrentHashMap<>();
-            count = new ConcurrentHashMap<>();
+        if (!isOpen(memory)) {
+            throw new IOException(CLOSED);
         }
-        openMemory(memory);
         int imported = databases.get(memory).storeTMX(tmx, project, client, subject);
-        closeMemory(memory);
         Files.deleteIfExists(new File(tmx).toPath());
         return imported;
     }
@@ -90,10 +91,9 @@ public class TmManager {
         Utils.removeDir(new File(memoriesFolder, memory));
     }
 
-    public static void closeMemory(String memory) throws SQLException {
-        if (databases == null) {
-            databases = new ConcurrentHashMap<>();
-            count = new ConcurrentHashMap<>();
+    public static void close(String memory) throws SQLException, IOException {
+        if (!isOpen(memory)) {
+            throw new IOException(CLOSED);
         }
         if (databases.containsKey(memory)) {
             int users = count.get(memory);
@@ -108,20 +108,20 @@ public class TmManager {
         }
     }
 
-    public static String exportMemory(String memory, String name) throws SQLException, IOException {
-        openMemory(memory);
-        InternalDatabase engine = databases.get(memory);
+    public static String exportMemory(String memory, String name, Set<String> languages, String srcLang) throws SQLException, IOException {
+        if (!isOpen(memory)) {
+            throw new IOException(CLOSED);
+        }
         File tempFolder = new File(RemoteTM.getWorkFolder(), "tmp");
         File tmx = new File(tempFolder, name + ".tmx");
         if (tmx.exists()) {
             Files.delete(tmx.toPath());
         }
-        engine.exportMemory(tmx.getAbsolutePath(), engine.getAllLanguages(), "*all*");
-        closeMemory(memory);
+        databases.get(memory).exportMemory(tmx.getAbsolutePath(), languages, srcLang);
         return tmx.getName();
     }
 
-    private static void openMemory(String memory) throws SQLException, IOException {
+    public static void openMemory(String memory) throws SQLException, IOException {
         if (databases == null) {
             databases = new ConcurrentHashMap<>();
             count = new ConcurrentHashMap<>();
@@ -134,7 +134,7 @@ public class TmManager {
         count.put(memory, count.get(memory) + 1);
     }
 
-    public static void closeMemories() throws SQLException {
+    public static void closeMemories() throws SQLException, IOException {
         if (databases == null) {
             databases = new ConcurrentHashMap<>();
             count = new ConcurrentHashMap<>();
@@ -142,7 +142,90 @@ public class TmManager {
         Set<String> keys = databases.keySet();
         Iterator<String> it = keys.iterator();
         while (it.hasNext()) {
-            closeMemory(it.next());
+            close(it.next());
         }
+    }
+
+    public static Set<String> getAllClients(String memory) throws SQLException, IOException {
+        if (!isOpen(memory)) {
+            throw new IOException(CLOSED);
+        }
+        return databases.get(memory).getAllClients();
+    }
+
+    public static Set<String> getAllLanguages(String memory) throws SQLException, IOException {
+        if (!isOpen(memory)) {
+            throw new IOException(CLOSED);
+        }
+        return databases.get(memory).getAllLanguages();
+    }
+
+    public static Set<String> getAllProjects(String memory) throws IOException {
+        if (!isOpen(memory)) {
+            throw new IOException(CLOSED);
+        }
+        return databases.get(memory).getAllProjects();
+    }
+
+    public static Set<String> getAllSubjects(String memory) throws IOException {
+        if (!isOpen(memory)) {
+            throw new IOException(CLOSED);
+        }
+        return databases.get(memory).getAllSubjects();
+    }
+
+    public static void storeTu(String memory, Element tu) throws SQLException, IOException {
+        if (!isOpen(memory)) {
+            throw new IOException(CLOSED);
+        }
+        databases.get(memory).storeTu(tu);
+    }
+
+    public static Element getTu(String memory, String tuid)
+            throws IOException, SQLException, SAXException, ParserConfigurationException {
+        if (!isOpen(memory)) {
+            throw new IOException(CLOSED);
+        }
+        return databases.get(memory).getTu(tuid);
+    }
+
+    public static void removeTu(String memory, String tuid) throws IOException, SQLException {
+        if (!isOpen(memory)) {
+            throw new IOException(CLOSED);
+        }
+        databases.get(memory).removeTu(tuid);
+    }
+
+    public static void commit(String memory) throws SQLException, IOException {
+        if (!isOpen(memory)) {
+            throw new IOException(CLOSED);
+        }
+        databases.get(memory).commit();
+    }
+
+    public static List<Match> searchTranslation(String memory, String searchStr, String srcLang, String tgtLang,
+            int similarity, boolean caseSensitive)
+            throws IOException, SAXException, ParserConfigurationException, SQLException {
+        if (!isOpen(memory)) {
+            throw new IOException(CLOSED);
+        }
+        return databases.get(memory).searchTranslation(searchStr, srcLang, tgtLang, similarity, caseSensitive);
+    }
+
+    public static List<Element> searchAll(String memory, String searchStr, String srcLang, int similarity,
+            boolean caseSensitive) throws IOException, SAXException, ParserConfigurationException, SQLException {
+        if (!isOpen(memory)) {
+            throw new IOException(CLOSED);
+        }
+        return databases.get(memory).searchAll(searchStr, srcLang, similarity, caseSensitive);
+    }
+
+    public static List<Element> concordanceSearch(String memory, String searchStr, String srcLang, int limit,
+            boolean isRegexp, boolean caseSensitive)
+            throws SQLException, SAXException, IOException, ParserConfigurationException {
+        if (!isOpen(memory)) {
+            throw new IOException(CLOSED);
+        }
+        return databases.get(memory).concordanceSearch(searchStr, srcLang, limit, isRegexp, caseSensitive);
     }
 }
